+++ projects/tests/async-hooks.spec.ts
@@ -148,6 +148,7 @@ describe('test with remult within get user & init request', () => {
     )
     const api = remultExpress({
       dataProvider: mem,
+      entities: [],
       initApi: async (req) => {
         await repo(t).insert({ id: '1' })
       },
@@ -175,6 +176,7 @@ describe('test default data provider based in init api server', () => {
     )
     const api = remultExpress({
       dataProvider: mem,
+      entities: [],
       initApi: async (req) => {
         await repo(t).insert({ id: '1' })
       },
@@ -188,5 +190,58 @@ describe('test default data provider based in init api server', () => {
   afterEach(() => {
     RemultAsyncLocalStorage.disable()
     remultStatic.asyncContext = new RemultAsyncLocalStorage(undefined)
+  })
+})
+describe('errors with withRemult', () => {
+  it('api with remult provides sensible error', async () => {
+    const api = remultExpress({
+      initApi: async (req) => {},
+      entities: [],
+    })
+    let error = false
+    try {
+      await api.withRemultAsync(undefined, async () => {
+        throw Error('the error')
+      })
+    } catch (err) {
+      error = true
+      expect(err).toMatchInlineSnapshot(`[Error: the error]`)
+    }
+    expect(error).toBe(true)
+  })
+  it('api with remult provides sensible error', async () => {
+    const api = remultExpress({
+      initApi: async (req) => {},
+      entities: [],
+    })
+    let error = false
+    try {
+      await api.withRemultAsync({} as any, async () => {
+        throw Error('the error')
+      })
+    } catch (err) {
+      error = true
+      expect(err).toMatchInlineSnapshot(`[Error: the error]`)
+    }
+    expect(error).toBe(true)
+  })
+  it('api with remult provides sensible error in get user', async () => {
+    const api = remultExpress({
+      getUser: async (req) => {
+        throw Error('get user error')
+      },
+      initApi: async (req) => {},
+      entities: [],
+    })
+    let error = false
+    try {
+      await api.withRemultAsync({} as any, async () => {
+        throw Error('the error')
+      })
+    } catch (err) {
+      error = true
+      expect(err).toMatchInlineSnapshot(`[Error: get user error]`)
+    }
+    expect(error).toBe(true)
   })
 })